// generated/src/app/api-models/smparalleloperation-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmparalleloperationS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  paralleloperationName: string;
  paralleloperationDescription: string;
  paralleloperationParallel: number;
  builtinoperationBuiltinid: number;
}